load-testing/requirements.txt
locust>=2.17.0
websockets>=11.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
